requirement.txt
PyJWT
python-decouple
fastapi
uvicorn